requirements.txt
-f https://pypi.research.pdx.edu/dist/
psu.oit.wdt.emcee>=1.0.0.dev0

